Move types to appropriate libs, other adjustments
Per title
